refactor: Created a separate file for renameDirectory
