+++ packages/files/tests/files/rename-directory/edge-case-no-match.test.js
@@ -1,0 +1,14 @@
+import { assert, test } from '@codemod-utils/tests';
+
+import { renameDirectory } from '../../../src/index.js';
+
+test('files | rename-directory > edge case (no match)', function () {
+  const oldFilePath = 'addon';
+
+  const newFilePath = renameDirectory(oldFilePath, {
+    from: 'addon',
+    to: 'new-location/src',
+  });
+
+  assert.strictEqual(newFilePath, 'addon');
+});
